requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiopg==1.4.0
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
black==23.7.0
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
Cython==3.0.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
isort==5.12.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.7
pycares==4.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
SQLAlchemy==1.4.49
typing_extensions==4.7.1
yarl==1.9.2
